Move fixtures to `jupyter_server_ydoc.pytest_plugin`
