+++ tests/test_app.py
@@ -4,9 +4,8 @@
 from __future__ import annotations
 
 import pytest
+from jupyter_server_ydoc.pytest_plugin import rtc_create_SQLite_store_factory
 from jupyter_server_ydoc.stores import SQLiteYStore, TempFileYStore
-
-from .conftest import rtc_create_SQLite_store_factory
 
 
 def test_default_settings(jp_serverapp):
